requirements.txt
aiohttp==3.6.2
astroid==2.2.5
async-timeout==3.0.1
attrs==20.1.0
certifi==2019.9.11
chardet==3.0.4
discord.py==1.4.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
multidict==4.7.6
pylint==2.3.1
python-dotenv==0.14.0
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.6
virtualenv==16.7.2
wrapt==1.11.2
yarl==1.5.1
